Improve Histogram Memory Usage in Group-By Case
Currently, an array of histogram objects per group-by-id is kept. This
is inefficient for two major reasons:

1. there is a large # of java objects and incur overhead. Inside
the histogram implementation, collections like LongBigArray or
IntBigArray are used which have minimize sizes in the 100-200+ byte
ranges. Histograms with a single item, or a few, still incur this
penalty.

2. across group-by-ids, there may be the same value multiple times.

* anecdotal analysis showed we do have group-by histograms that have
few values
** after implementation, we found with yourkit that use cases that
were holding 1 gb of ram dropped to 45mb of ram.
*** profiling with JMH showed that the new version was equal to or
better than the old version

what:

The solution here is to build a single object that includes the
group-by-id in the key within an internal hash table. There are
many details to support the appearance of a single histogram
per group-by table both for transfer when doing intermediate
combining, as well as final output. See GroupedTypedHistogram for
the details. In short, this reduces the # of java objects to a constant
number. Values are also stored once in another custom hash
implementation.

We also found that changing the probe function from linear to
"sum of square" (sort of super-quadratic) performed better testing
with JMH.

This has also introduced a new config variable, or rather used one
from a previous commit, that allows legacy (existing behavior) or
the new one. We default to the new behavior, but allow rollback
to the old behavior in case of some unexpected failure.

Future work for readability, private inner classes were introduced.
These don't appear to have affected perf (per JMH), but JFR confirms
they aren't elided by the JVM. This is an opportunity to improve perf
(causes 4.5% of gc pressure in some testing).